Restructure folders, move components/assets out of /app.
