+++ components/assets/icons/Coffee.jsx
